feat: group chat

+++ components/Event/EventModal/Host/InviteView.tsx
@@ -1,21 +1,23 @@
 import ModalFooter from '@/components/Common/ModalFooter'
 import Loader from '@/components/State/Loader'
 import getDateTime from '@/lib/getDateTime'
+import getName from '@/lib/getName'
 import EventModalView from '@/types/EventModalView'
+import { Guest } from '@/types/Guest'
 import InviteRevokedProps from '@/types/email/InviteRevokedProps'
-import { Event, Invite } from '@prisma/client'
+import { Event } from '@prisma/client'
 import { Dispatch, SetStateAction } from 'react'
 
 export default function InviteView({
   event,
-  invite,
+  guest,
   setView,
   setIsLoading,
   isLoading,
   onDelete,
 }: {
   event: Event
-  invite: Invite | undefined
+  guest: Guest | undefined
   setView: Dispatch<SetStateAction<EventModalView>>
   setIsLoading: Dispatch<SetStateAction<boolean>>
   isLoading: boolean
@@ -25,23 +27,25 @@ export default function InviteView({
     try {
       setIsLoading(true)
 
-      const { status } = await fetch(`/api/invite?id=${invite?.id}`, {
-        method: 'DELETE',
-      })
-      console.log('delete status:', status)
+      if (guest) {
+        const { status } = await fetch(`/api/invite?id=${guest.id}`, {
+          method: 'DELETE',
+        })
+        console.log('delete status:', status)
 
-      const emailRes = await fetch(`/api/email?to=${invite?.email}&template=invite-revoked`, {
-        method: 'POST',
-        body: JSON.stringify({
-          title: event.title,
-          username: invite?.firstName,
-        } as InviteRevokedProps),
-      })
+        const emailRes = await fetch(`/api/email?to=${guest.email}&template=invite-revoked`, {
+          method: 'POST',
+          body: JSON.stringify({
+            title: event.title,
+            username: getName(guest),
+          } as InviteRevokedProps),
+        })
 
-      console.log('email res:', await emailRes.json())
+        console.log('email res:', await emailRes.json())
 
-      onDelete()
-      setView(EventModalView.INFO)
+        onDelete()
+        setView(EventModalView.INFO)
+      }
     } catch (err) {
       console.error(err)
     } finally {
@@ -53,21 +57,19 @@ export default function InviteView({
     <>
       <div className="p-4 sm:px-7 sm:py-6">
         <div className="pb-8">
-          <h1 className="text-xl font-medium sm:text-2xl">
-            {invite?.firstName} {invite?.lastName}
-          </h1>
+          <h1 className="text-xl font-medium sm:text-2xl">{getName(guest as Guest)}</h1>
           <p className="text-sm text-gray-500 sm:text-base">
-            {invite?.scannedAt ? `Ticket scanned ${getDateTime(new Date(invite.scannedAt))}` : 'Ticket not yet scanned'}
+            {guest?.scannedAt ? `Ticket scanned ${getDateTime(new Date(guest.scannedAt))}` : 'Ticket not yet scanned'}
           </p>
         </div>
         <ul className="flex flex-col space-y-2 sm:space-y-0">
           <li className="sm:flex">
             <p className="w-1/5 text-sm font-semibold uppercase text-gray-500">Email</p>
-            <p className="w-4/5 text-gray-800">{invite?.email}</p>
+            <p className="w-4/5 text-gray-800">{guest?.email}</p>
           </li>
           <li className="sm:flex">
             <p className="w-1/5 text-sm font-semibold uppercase text-gray-500">Accepted</p>
-            <p className="w-4/5 text-gray-800">{getDateTime(new Date(invite?.acceptedAt as Date))}</p>
+            <p className="w-4/5 text-gray-800">{getDateTime(new Date(guest?.acceptedAt as Date))}</p>
           </li>
         </ul>
       </div>
@@ -83,7 +85,7 @@ export default function InviteView({
           <button
             className="rounded-lg bg-gray-600 px-5 py-2.5 text-sm font-semibold text-gray-200 shadow-sm transition-colors duration-200 hover:bg-gray-700 hover:text-gray-100 disabled:opacity-50 hover:disabled:bg-gray-600 hover:disabled:text-gray-200"
             onClick={deleteInvite}
-            disabled={invite?.scannedAt ? true : false}
+            disabled={guest?.scannedAt ? true : false}
           >
             Revoke Invite
           </button>
